First stab at installer for MySQL (tested on RedHat 6 and SuSE 10 only).
This is a first stab, and not tested on Debian/Ubuntu systems (won't work there anyway due to library issues, will chat with PM about this).  This is primarily to get something that basically works (at least on tested platforms), and get to a point where it can easily be evolved as needed.

+++ source/code/providers/MySQL_ServerStatistics_Class_Provider.cpp
@@ -20,7 +20,7 @@ static void ComputeRatio(float numerator, float denominator, uint8_t& ratio, boo
 
     if ( denominator ) // Avoid divide by zero
     {
-        ratio = (numerator / denominator) * 100.0;
+        ratio = static_cast<uint8_t> ((numerator / denominator) * 100.0);
     }
 }
